refactor(Scheduler): move *Scheduler to main package
