+++ core/src/main/java/io/github/zero88/schedulerx/CronScheduler.java
@@ -1,9 +1,9 @@
-package io.github.zero88.schedulerx.trigger;
+package io.github.zero88.schedulerx;
 
 import org.jetbrains.annotations.NotNull;
 
-import io.github.zero88.schedulerx.Scheduler;
-import io.github.zero88.schedulerx.trigger.CronSchedulerImpl.CronSchedulerBuilderImpl;
+import io.github.zero88.schedulerx.CronSchedulerImpl.CronSchedulerBuilderImpl;
+import io.github.zero88.schedulerx.trigger.CronTrigger;
 import io.vertx.codegen.annotations.GenIgnore;
 import io.vertx.codegen.annotations.VertxGen;
 import io.vertx.core.Vertx;
