+++ core/src/main/java/io/github/zero88/schedulerx/ExecutionContext.java
@@ -7,7 +7,6 @@ import org.jetbrains.annotations.Nullable;
 
 import io.github.zero88.schedulerx.trigger.TriggerContext;
 import io.vertx.codegen.annotations.GenIgnore;
-import io.vertx.codegen.annotations.VertxGen;
 import io.vertx.core.Vertx;
 
 /**
@@ -17,7 +16,6 @@ import io.vertx.core.Vertx;
  * @apiNote This interface is renamed from {@code TaskExecutionContext} since {@code 2.0.0}
  * @since 1.0.0
  */
-@VertxGen
 public interface ExecutionContext<OUT> extends HasVertx {
 
     @NotNull Vertx vertx();
@@ -35,7 +33,6 @@ public interface ExecutionContext<OUT> extends HasVertx {
      * @see TriggerContext
      * @since 2.0.0
      */
-    @GenIgnore(GenIgnore.PERMITTED_TYPE)
     @NotNull TriggerContext triggerContext();
 
     /**
