+++ core/src/main/java/io/github/zero88/schedulerx/Scheduler.java
@@ -7,7 +7,9 @@ import io.vertx.core.WorkerExecutor;
  * A scheduler schedules a job to run based on a particular trigger.
  *
  * @apiNote This interface is renamed from {@code TriggerTaskExecutor} since {@code 2.0.0}
- * @see Trigger
+ * @see IntervalScheduler
+ * @see CronScheduler
+ * @see EventScheduler
  * @since 2.0.0
  */
 public interface Scheduler<TRIGGER extends Trigger> extends HasTrigger<TRIGGER>, HasVertx {
