+++ core/src/test/java/io/github/zero88/schedulerx/EventSchedulerTest.java
@@ -1,4 +1,4 @@
-package io.github.zero88.schedulerx.trigger;
+package io.github.zero88.schedulerx;
 
 import static org.junit.jupiter.params.provider.Arguments.arguments;
 
@@ -18,12 +18,8 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-import io.github.zero88.schedulerx.ExecutionResult;
-import io.github.zero88.schedulerx.Job;
-import io.github.zero88.schedulerx.NoopJob;
-import io.github.zero88.schedulerx.SchedulingAsserter;
-import io.github.zero88.schedulerx.SchedulingMonitor;
-import io.github.zero88.schedulerx.TestUtils;
+import io.github.zero88.schedulerx.trigger.EventTrigger;
+import io.github.zero88.schedulerx.trigger.TriggerCondition;
 import io.github.zero88.schedulerx.trigger.predicate.EventTriggerPredicate;
 import io.vertx.core.MultiMap;
 import io.vertx.core.Vertx;
