+++ core/src/test/java/io/github/zero88/schedulerx/custom/HttpClientJobTest.java
@@ -7,9 +7,9 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
 import io.github.zero88.schedulerx.ExecutionResult;
+import io.github.zero88.schedulerx.IntervalScheduler;
 import io.github.zero88.schedulerx.JobData;
 import io.github.zero88.schedulerx.SchedulingAsserter;
-import io.github.zero88.schedulerx.trigger.IntervalScheduler;
 import io.github.zero88.schedulerx.trigger.IntervalTrigger;
 import io.vertx.core.Vertx;
 import io.vertx.core.json.JsonObject;
